producer/requirements.txt
confluent-kafka
kafka-python-ng